Stream API [#106993] - исправил решение

+++ chapter_003/test/java/ru/job4j/arraysum/ArraySumTest.java
@@ -13,7 +13,7 @@ public class ArraySumTest {
     public void whenAddingUserAccount() {
         ArraySum arraySum = new ArraySum();
         Integer[] array = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
-        Double result =arraySum.sum(array);
-        assertThat(result, is(220D));
+        int result = arraySum.sum(array);
+        assertThat(result, is(220));
     }
 }
